Add simplified V1 algorithm without DA
